Set hwservicemanager.disabled to false even when it's enabled
This is a reland of d97e72a4d1aae4cd5e21c8278a4b618874994455

Currently libhwbinder waits for hwservicemanager.ready to be true with
1-sec timeout in a loop. However, On devices with hwservicemanager
disabled, .ready is never set, and that causes a race that forces
libhwbinder to wait for the timeout. This unnecessarily slows down the
boot time. This CL fixes this issue.

To prevent this issue, this change makes hwservicemanager always
initialize hwservicemanager.disabled even when it's false. That allows
libhwbinder to wait for hwservicemanager.disabled first on all devices,
instead of hwservicemanager.ready. So libhwbinder will no longer be
forced to be blocked for 1 second. This change improves boot time on
some devices by up to 300ms.

There is no guarantee libhwbinder and hwservicemanager are updated
together. Actually, older hwservicemanager can run with newer
libhwbinder and vice versa. So, we set
`hwservicemanager.always_sets_disabled` system property to indicate the
new behavior.

Bug: 419725480
Test: CF boots. Confirmed that .disabled is checked by additional log
Test: Confirmed no timeout messages appear in logcat on CF
Test: CF boots with ToT libhwbinder as well
Change-Id: Ic66b9e19ff737651384ac50a23730276720e91a0