Explicitly set hwservicemanager.disabled to false
Currently libhwbinder waits for hwservicemanager.ready with timeout of 1
second in a loop. However, On devices with hwservicemanager disabled,
this property is never set, and that causes a race that forces
libhwbinder to wait for the timeout. This can be fixed by letting
libhwbinder wait for hwservicemanager.disabled first, but
hwservicemanager.disabled is never set on devices with hwservicemanager
enabled.

To prevent this issue, this change makes hwservicemanager always
initialize hwservicemanager.disabled even when it's false. That allows
libhwbinder to wait for hwservicemanager.disabled first on all devics,
instead of hwservicemanager.ready. So libhwbinder will no longer be
forced to be blocked for 1 second.

Bug: 419725480
Test: manual
Change-Id: Ia39940ee78ffd23951acbfc28978bb3fc2a35a8a